refactor: extract cli/commands/
