+++ src/incus-app-container-files/cli/commands/list-app-containers.ts
@@ -1,0 +1,3 @@
+export function listAppContainers(format: string): void {
+  console.log(`Listing containers as ${format}...`);
+}
